+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is *inspired* by `Keep a Changelog <http://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 
-`v0.6.0`_ - 00-Unreleased-0000
+`v0.6.0`_ - 4-September-2021
 ------------------------------
 Added
 +++++